sentiment_analysis/requirements.txt
django==4.2.21
djangorestframework==3.15.2
django-filter==24.3
python-decouple==3.8
psycopg2-binary==2.9.10
tweepy==4.14.0
praw==7.8.1
textblob==0.18.0.post0

whitenoise==6.7.0
gunicorn==23.0.0
